chore: rename lms service to cms
